change ID fields to bigint

+++ internal/db/models.go
@@ -4,27 +4,27 @@ import "time"
 
 type BinanceOrderBook struct {
 	ID uint 			`gorm:"primary_key"`
-	MarketID uint
+	MarketID uint64
 	Time time.Time		`gorm:"primary_key"`
 }
 
 type BinanceOrder struct {
-	ID uint 			`gorm:"primary_key"`
+	ID uint64 			`gorm:"primary_key"`
 	Time time.Time		`gorm:"primary_key"`
-	OrderbookID uint
+	OrderbookID uint64
 	Rate float64
 	Quantity float64
 }
 
 type BinanceMarket struct {
-	ID uint 			`gorm:"primary_key"`
+	ID uint64 			`gorm:"primary_key"`
 	Ticker string		`gorm:"unique_index:time_idx_market"`
 	Quote string		`gorm:"unique_index:time_idx_market"`
 }
 
 type BinanceTicker struct {
-	ID  uint 			`gorm:"primary_key"`
-	MarketID uint
+	ID  uint64 			`gorm:"primary_key"`
+	MarketID uint64
 	Price float64
 	Time time.Time		`gorm:"primary_key"`
 }
